requirements.txt
kagglehub
Pillow
dotmap
matplotlib
clean-fid
lpips
wandb
